fixing move smooth to use solve from a1

+++ openrb_dynamixel_v22/vertical_kinematics.h
@@ -63,4 +63,4 @@ private:
 // -------------------------------------------------------------------
 
 extern VerticalKinematics kin;
-extern float l_mm;
+extern double l_mm;
